functions/picture/requirements.txt
black==22.3.0
Flask==2.1.2
functions-framework==3.0.0
google-cloud-storage==2.3.0
pg8000==1.26.1
pylint==2.13.9
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
